Update listings css
