+++ client/src/app/main/Main.css
@@ -1,9 +1,6 @@
 .main {
   width: 80%;
   margin: 0 auto;
-  /* background-color: azure; */
-  /* border-radius: 10px; */
-  /* box-shadow: 0px 0px 3px 3px gray; */
 }
 
 .main-background-image {
@@ -17,4 +14,8 @@
   background-repeat: repeat;
   background-size: contain;
   mask-image: linear-gradient(to bottom, rgba(0, 0, 0, 1), 75%, rgba(0, 0, 0, 0));
+}
+
+.page {
+  padding: 20px;
 }
