+++ client/src/listings/ListingList.js
@@ -1,52 +1,26 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
-import Typography from '@material-ui/core/Typography';
-import Accordion from '@material-ui/core/Accordion';
-import AccordionSummary from '@material-ui/core/AccordionSummary';
-import AccordionDetails from '@material-ui/core/AccordionDetails';
+import {
+  Grid,
+  Paper,
+  Typography,
+  Accordion,
+  AccordionSummary,
+  AccordionDetails,
+  Button,
+  Card,
+} from '@material-ui/core';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import Button from '@material-ui/core/Button';
 import Rating from '@material-ui/lab/Rating';
-import '../shared/List.css';
 import './Listings.css';
-// import Ratings from '../shared/ratings';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-    maxWidth: '1000px',
-    paddingTop: '25px',
-  },
-  paper: {
-    padding: theme.spacing(2),
-    margin: 0,
-  },
-  img: {
-    maxWidth: '250px',
-    maxHeight: '250px',
-    borderRadius: '5%',
-  },
-  accordionImg: {
-    marginRight: '25px',
-    maxWidth: '75px',
-    maxHeight: '75px',
-    borderRadius: '50%',
-  },
-}));
 
 export default function ListingList({ listings }) {
-  const classes = useStyles();
-
   return (
-    <div className={classes.root}>
+    <Card>
       {listings.map((listing) => (
         <Accordion key={listing.id}>
           <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls='panel1a-content' id='panel1a-header'>
             <Grid container spacing={2} direction='row' alignItems='center' justify='space-between'>
               <Grid item container spacing={2} direction='row' alignItems='center' justify='flex-start' xs={10}>
-                <img className={classes.accordionImg} alt='Sample' src={require('../shared/img.png').default} />
+                <img className='listing-accordion-image' alt='Sample' src={require('../shared/img.png').default} />
                 <Grid item>
                   <Typography variant='h5'>{listing.title}</Typography>
                 </Grid>
@@ -57,13 +31,7 @@ export default function ListingList({ listings }) {
               <Grid item container spacing={2} direction='row' alignItems='center' justify='flex-end' xs={2}>
                 <Grid item>
                   {/* <Ratings Ratings={Ratings}></Ratings> */}
-                  <Rating
-                    name='rating'
-                    value={listing.ratingAverage}
-                    defaultValue={5}
-                    precision={0.5}
-                    readOnly
-                  />
+                  <Rating name='rating' value={listing.ratingAverage} defaultValue={5} precision={0.5} readOnly />
                 </Grid>
               </Grid>
             </Grid>
@@ -71,30 +39,30 @@ export default function ListingList({ listings }) {
           <AccordionDetails>
             <Grid container spacing={2}>
               <Grid item>
-                <img className={classes.img} alt='Sample' src={require('../shared/img.png').default} />
+                <img className='listing-image' alt='Sample' src={require('../shared/img.png').default} />
               </Grid>
               <Grid item xs container>
                 <Grid item xs container direction='column' spacing={1}>
                   <Grid item xs={12} sm container spacing={1} alignItems='baseline' justify='space-between'>
                     <Grid item xs>
-                      <Paper className={classes.paper}>
+                      <Paper className='listing-item'>
                         <Typography variant='body1'>{listing.description}</Typography>
                       </Paper>
                     </Grid>
                   </Grid>
                   <Grid item xs sm container spacing={1}>
                     <Grid item xs>
-                      <Paper className={classes.paper}>
+                      <Paper className='listing-item'>
                         <Typography variant='body2'>{listing.website}</Typography>
                       </Paper>
                     </Grid>
                     <Grid item xs>
-                      <Paper className={classes.paper}>
+                      <Paper className='listing-item'>
                         <Typography variant='body2'>{listing.email}</Typography>
                       </Paper>
                     </Grid>
                     <Grid item xs={3}>
-                      <Paper className={classes.paper}>
+                      <Paper className='listing-item'>
                         <Typography variant='body2'>{listing.phone}</Typography>
                       </Paper>
                     </Grid>
@@ -124,7 +92,7 @@ export default function ListingList({ listings }) {
               </Grid>
               <Grid container spacing={0}>
                 <Grid item xs={12}>
-                  <Paper className={classes.paper}>
+                  <Paper className='listing-item'>
                     <Typography variant='body2'>
                       Published: {listing.creationDate.slice(0, 10)} by {listing.organizationName}
                     </Typography>
@@ -135,6 +103,6 @@ export default function ListingList({ listings }) {
           </AccordionDetails>
         </Accordion>
       ))}
-    </div>
+    </Card>
   );
 }
