+++ client/src/listings/Listings.css
@@ -1,26 +1,51 @@
-.listings-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 80%;
-  border: 1px solid #f2f2f2;
-  border-radius: 20px;
-  padding: 30px;
-  margin-top: 20px;
-}
-
 .listing-select {
   width: 450px;
   margin-left: 8px;
 }
 
+.listing-item {
+  padding: 10px;
+  margin: 0;
+}
+
+.listing-image {
+  max-width: 250px;
+  max-height: 250px;
+  border-radius: 5%;
+}
+
+.listing-accordion-image {
+  margin-right: 25px;
+  max-width: 75px;
+  max-height: 75px;
+  border-radius: 50%;
+}
+
+.listing-title {
+  cursor: pointer;
+}
+
+.listing-row {
+  position: relative;
+  border: solid 1px gray;
+  padding: 15px;
+  margin: 1px;
+}
+
+.listing-row:hover {
+  box-shadow: 3px 3px gray;
+  top: -2px;
+  left: -2px;
+}
+
 .MuiPaper-root {
   text-align: left;
-
 }
-.ratingAlign{
+
+.ratingAlign {
   align-content: right;
 }
+
 .MuiAccordionSummary-content.Mui-expanded {
   margin: 0;
 }
